finished fonts and gameplay, hoisted renderer to separate thing, incoming opengl layer

+++ src/renderer_software.h
@@ -7,8 +7,8 @@ typedef struct software_renderer
 {
   // This should be a pool?
   u32*     Buffer;
-  u16      Width;
-  u16      Height;
+  u32      Width;
+  u32      Height;
 } software_renderer;
 
 #endif
